+++ Spartane.Web/Areas/Frontal/Models/Tipo_de_Equipo_TacticoGridModel.cs
@@ -8,7 +8,7 @@ namespace Spartane.Web.Areas.Frontal.Models
 {
     public class Tipo_de_Equipo_TacticoGridModel
     {
-        public short Clave { get; set; }
+        public int Clave { get; set; }
         public string Descripcion { get; set; }
         
     }
